Added better error handling for the Boa tester (#1984)
Trying to fix the issue in #1982, I noticed that we didn't have a proper error handling for the boa tester.

This adds the `anyhow` dependency to the tester, which makes it much easier to handle errors and bubble them up with attached context. Thanks to this I was able to easily find out the issue, and I think it could be useful to have it. It gives errors such as this one:

```
Error: could not read the suite test
    caused by: error reading sub-suite ./test262/test/built-ins
    caused by: error reading sub-suite ./test262/test/built-ins/ShadowRealm
    caused by: error reading sub-suite ./test262/test/built-ins/ShadowRealm/WrappedFunction
    caused by: error reading test ./test262/test/built-ins/ShadowRealm/WrappedFunction/throws-typeerror-on-revoked-proxy.js
    caused by: while scanning a block scalar, found a tab character where an indentation space is expected at line 4 column 3
    caused by: while scanning a block scalar, found a tab character where an indentation space is expected at line 4 column 3
```